+++ src/core/gameObjects/LiveObjects/EnemyObjects/EChaser.java
@@ -7,11 +7,11 @@ import core.gameObjects.LiveObjects.LOEnemy;
  * Created by Andrew on 13/08/2016.
  */
 public class EChaser extends LOEnemy {
-    public EChaser(Game game, int x, int y, int hp) {
+    public EChaser(Game game, Double x, Double y, Double hp) {
         super(game, x, y, hp);
     }
 
-    public void goTo(int x, int y){
+    public void goTo(Double x, Double y){
         turnToLookAt(x, y);
         calcVelocity();
     }
