+++ src/core/gameObjects/LiveObjects/EnemyObjects/EShooter.java
@@ -10,14 +10,14 @@ public class EShooter extends LOEnemy {
 
     public static final int DEFAULT_SHOOTER_ACTION_INTERVAL = 1000;
 
-    public EShooter(Game game, int x, int y, int hp) {
+    public EShooter(Game game, Double x, Double y, Double hp) {
         super(game, x, y, hp);
 
         speed = 0;
         actionInterval = DEFAULT_SHOOTER_ACTION_INTERVAL;
     }
 
-    public void shootAt(int x, int y){
+    public void shootAt(Double x, Double y){
         turnToLookAt(x, y);
         tryToFire();
     }
