+++ src/core/gameObjects/LiveObjects/LOEnemy.java
@@ -8,11 +8,11 @@ import core.gameObjects.GOLiveObject;
  */
 public abstract class LOEnemy extends GOLiveObject {
     public static final String SPRITES_ENEMY_GIF = "sprites/enemy.gif";
-    public static final int DEFAULT_ENEMY_MOVE_SPEED = 75;
-    public static final int DEFAULT_ENEMY_HP = 1;
+    public static final Double DEFAULT_ENEMY_MOVE_SPEED = 75.0;
+    public static final Double DEFAULT_ENEMY_HP = 1.0;
 
 
-    public LOEnemy(Game game, int x, int y, int hp) {
+    public LOEnemy(Game game, Double x, Double y, Double hp) {
         super(game, SPRITES_ENEMY_GIF, x, y, hp);
 
         speed = DEFAULT_ENEMY_MOVE_SPEED;
