+++ src/core/gameObjects/LiveObjects/LOPlayer.java
@@ -9,10 +9,10 @@ import core.gameObjects.GOLiveObject;
  */
 public class LOPlayer extends GOLiveObject {
     public static final String SPRITES_SHIP_GIF = "sprites/player.gif";
-    public static final int DEFAULT_SHIP_MOVE_SPEED = 200;
-    public static final int DEFAULT_PLAYER_HP = 3;
+    public static final Double DEFAULT_SHIP_MOVE_SPEED = 200.0;
+    public static final Double DEFAULT_PLAYER_HP = 3.0;
 
-    public LOPlayer(Game game, int x, int y) {
+    public LOPlayer(Game game, Double x, Double y) {
         super(game, SPRITES_SHIP_GIF, x, y, DEFAULT_PLAYER_HP);
 
         speed = DEFAULT_SHIP_MOVE_SPEED;
